+++ RELEASES.md
@@ -16,7 +16,8 @@
 
 ### 0.2.0 Removed
 
-- Nothing yet
+- various internal types are no longer public
+  - if you needed one of these, please file an issue!
 
 ## 0.1.0